Read-WriteRESTAPI/requirements.txt
django==2.1.11
djangorestframework==3.9.1